+++ src/main/java/subway/controller/LineController.java
@@ -1,0 +1,67 @@
+package subway.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import subway.dto.LineRequest;
+import subway.dto.LineResponse;
+import subway.service.LineService;
+import subway.dto.SectionRequest;
+import subway.service.SectionService;
+
+
+import java.net.URI;
+import java.util.List;
+
+
+@RestController
+@RequestMapping("/lines")
+public class LineController {
+
+    private LineService lineService;
+    private SectionService sectionService;
+
+    public LineController(LineService lineService, SectionService sectionService){
+        this.lineService = lineService;
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping
+    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
+        LineResponse lineResponse = lineService.create(lineRequest);
+        return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<LineResponse>> showLines(){
+        return ResponseEntity.ok().body(lineService.showLines());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<LineResponse> showLine(@PathVariable Long id) {
+        return ResponseEntity.ok().body(lineService.showLine(id));
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity modifyLine(@PathVariable Long id, @RequestBody LineRequest lineRequest){
+        lineService.modify(id, lineRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity deleteLine(@PathVariable Long id) {
+        lineService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/{id}/sections")
+    public ResponseEntity addSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
+        sectionService.add(id, sectionRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/{id}/sections")
+    public ResponseEntity deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
+        sectionService.delete(id, stationId);
+        return ResponseEntity.ok().build();
+    }
+}
