+++ src/main/java/subway/dto/LineResponse.java
@@ -1,0 +1,39 @@
+package subway.dto;
+
+import subway.domain.Line;
+
+
+import java.util.List;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+    private int extraFare;
+    private List<StationResponse> stations;
+
+    public LineResponse () {}
+
+    public LineResponse(Line line, List<StationResponse> stations) {
+        this.id = line.getId();
+        this.name = line.getName();
+        this.color = line.getColor();
+        this.stations = stations;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+}
