--- src/main/java/subway/line/controller/LineController.java
@@ -1,67 +1,0 @@
-package subway.line.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.line.dto.LineRequest;
-import subway.line.dto.LineResponse;
-import subway.line.service.LineService;
-import subway.section.dto.SectionRequest;
-import subway.section.service.SectionService;
-
-
-import java.net.URI;
-import java.util.List;
-
-
-@RestController
-@RequestMapping("/lines")
-public class LineController {
-
-    private LineService lineService;
-    private SectionService sectionService;
-
-    public LineController(LineService lineService, SectionService sectionService){
-        this.lineService = lineService;
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping
-    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
-        LineResponse lineResponse = lineService.create(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<LineResponse>> showLines(){
-        return ResponseEntity.ok().body(lineService.showLines());
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<LineResponse> showLine(@PathVariable Long id) {
-        return ResponseEntity.ok().body(lineService.showLine(id));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity modifyLine(@PathVariable Long id, @RequestBody LineRequest lineRequest){
-        lineService.modify(id, lineRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity deleteLine(@PathVariable Long id) {
-        lineService.delete(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/{id}/sections")
-    public ResponseEntity addSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
-        sectionService.add(id, sectionRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/{id}/sections")
-    public ResponseEntity deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
-        sectionService.delete(id, stationId);
-        return ResponseEntity.ok().build();
-    }
-}
