--- src/main/java/subway/line/domain/Line.java
@@ -1,49 +1,0 @@
-package subway.line.domain;
-
-import java.util.Objects;
-
-public class Line {
-    public static final long HEAD = 0;
-    public static final long TAIL = -1;
-    public static final int INF = Integer.MAX_VALUE;
-    public static final String TERMINAL_NAME = "virtual";
-    private Long id;
-    private String name;
-    private String color;
-
-    public Line(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public Line(Long id, String name, String color) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Line line = (Line) o;
-        return Objects.equals(id, line.id) && Objects.equals(name, line.name) && Objects.equals(color, line.color);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, color);
-    }
-}
