--- src/main/java/subway/station/controller/StationController.java
@@ -1,39 +1,0 @@
-package subway.station.controller;
-
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.station.dto.StationRequest;
-import subway.station.dto.StationResponse;
-import subway.station.service.StationService;
-
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-
-    private StationService stationService;
-
-    public StationController(StationService stationService){
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        StationResponse stationResponse = stationService.createLine(stationRequest);
-        return ResponseEntity.created(URI.create("/stations/" + stationResponse.getId())).body(stationResponse);
-    }
-
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<StationResponse>> showStations() {
-        return ResponseEntity.ok().body(stationService.showStations());
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity deleteStation(@PathVariable Long id) {
-        stationService.deleteStation(id);
-        return ResponseEntity.noContent().build();
-    }
-}
