--- src/main/java/subway/station/dto/StationRequest.java
@@ -1,16 +1,0 @@
-package subway.station.dto;
-
-public class StationRequest {
-    private String name;
-
-    public StationRequest() {
-    }
-
-    public StationRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
