+++ src/test/java/subway/line/SectionAcceptanceTest.java
@@ -9,9 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.line.dto.LineResponse;
-import subway.section.dto.SectionRequest;
-import subway.station.dto.StationResponse;
+import subway.dto.LineResponse;
+import subway.dto.SectionRequest;
+import subway.dto.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
